Trying to do something useful

+++ src/excel.c
@@ -3,7 +3,7 @@
 void writeColumnNamesToWorksheet(char **columnNames, size_t numberOfColumns, 
                                                 lxw_worksheet *worksheet)
 {
-    for(int columnCount=0;columnCount<numberOfColumns;columnCount++)
+    for(size_t columnCount=0;columnCount<numberOfColumns;columnCount++)
     {
         worksheet_write_string(worksheet, 0, columnCount, columnNames[columnCount], NULL);    
         free(columnNames[columnCount]);
@@ -19,15 +19,16 @@ int generateExcelFile(char *fileName, char *columnNames,
     int numberOfColumnsInRow = 0;
     lxw_workbook  *workbook  = workbook_new(fileName);
     lxw_worksheet *worksheet = workbook_add_worksheet(workbook, NULL);
-    char **columns = splitString(columnNames, &numberOfColumnsInRow, "\t", headerColumnsNumber); // using numberOfColumnsInRow just to not make another variable
+    // Using numberOfColumnsInRow just to not make another variable
+    char **columns = splitString(columnNames, &numberOfColumnsInRow, "\t", headerColumnsNumber);
     writeColumnNamesToWorksheet(columns, headerColumnsNumber, worksheet);
 
     char **databaseBuffer = fetchall();
-    int databaseRowsCount = getBufferRowsCount(); 
+    size_t databaseRowsCount = getBufferRowsCount();
     for(size_t row = 0;row<databaseRowsCount;row++)
     {
         columns = splitString(databaseBuffer[row], &numberOfColumnsInRow, "\t", headerColumnsNumber);
-        for(size_t col = 0;col<numberOfColumnsInRow;col++)
+        for(int col = 0;col<numberOfColumnsInRow;col++)
         {
             worksheet_write_string(worksheet, row+1, col, columns[col], NULL);
         }
